add function to Rotate the rectangle and separate the BinaryTreeModel class into multiple classes
