+++ rotate.txt
@@ -1,0 +1,60 @@
+----------------------------------------------------------------------
+|                   |                   |                            |
+|                   |                   |                            |
+|                   |                   |                            |
+|                   |                   |                            |
+|         A         |         B         |              C             |
+|                   |                   |                            |
+|                   |                   |                            |
+|                   |                   |                            |
+|                   |                   |                            |
+----------------------------------------------------------------------
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                   E                  |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|              D              |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             ----------------------------------------
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                   F                  |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|                             |                                      |
+|-----------------------------|--------------------------------------|
