+++ src/Controller/TreePageController.java
@@ -7,7 +7,9 @@ import java.util.ArrayList;
 import javax.swing.JFrame;
 
 import Model.BinaryTreeModel;
+import Model.FileOperations;
 import Model.Node;
+import Model.RectangleOperations;
 import Views.Components.StaticMethods;
 import Views.Components.TextFieldNode;
 import Views.pages.RectangleView;
@@ -54,14 +56,16 @@ public class TreePageController {
             return;
         }
         initRoot();
+        FileOperations F=new FileOperations();
+        RectangleOperations R=new RectangleOperations();
         ArrayList<Node> response = model.convertToPaper();
         if (response == null) {
             StaticMethods.showMassage("The tree isn't fully grown yet ", frame, Type.INFO);
         } else if (response.isEmpty()) {
             StaticMethods.showMassage("Can't form rectangle", frame, Type.WARNING);
         } else {
-            char[][] rec=model.drawing(model.getRoot());
-            model.print2DArrayToFile(rec, "D:\\Projects\\Second Year Project\\Slicing Tree\\output.txt");
+            char[][] rec=R.drawing(model.getRoot());
+            F.print2DArrayToFile(rec, "D:\\Projects\\Second Year Project\\Slicing Tree\\output.txt");
             StaticMethods.showMassage("Successfuly convert the tree to rectangle", frame, Type.SUCCESS);
             rectangleView.addRectangles(response);
             ((CardLayout) frame.getContentPane().getLayout()).show(frame.getContentPane(), "Rectangle");
